Update README.md
For specificity: this is Newsletters resource, not Newsletter model. Perhaps might want to add a paragraph referencing/reminding that in the Model, by convention the class is Newletter/singular. But in app.py we are creating Newletters and NewsletterById resources. (I'm assuming that we can't create two distinct Newsletter classes and this is why we create NewsletterById resource instead of Newsletter resource.)